drivers: syscon: use uint32_t for register offsets
Use uint32_t to allow larger offsets in place of the existing uint16_t
configuration.

+++ drivers/syscon/syscon.c
@@ -35,7 +35,7 @@ static int syscon_generic_get_base(const struct device *dev, uintptr_t *addr)
 	return 0;
 }
 
-static int syscon_generic_read_nolock(const struct device *dev, uint16_t reg, uint32_t *val)
+static int syscon_generic_read_nolock(const struct device *dev, uint32_t reg, uint32_t *val)
 {
 	const struct syscon_generic_config *config = dev->config;
 	struct syscon_generic_data *data = dev->data;
@@ -68,7 +68,7 @@ static int syscon_generic_read_nolock(const struct device *dev, uint16_t reg, ui
 	return 0;
 }
 
-static int syscon_generic_write_nolock(const struct device *dev, uint16_t reg, uint32_t val)
+static int syscon_generic_write_nolock(const struct device *dev, uint32_t reg, uint32_t val)
 {
 	const struct syscon_generic_config *config = dev->config;
 	struct syscon_generic_data *data = dev->data;
@@ -97,7 +97,7 @@ static int syscon_generic_write_nolock(const struct device *dev, uint16_t reg, u
 	return 0;
 }
 
-static int syscon_generic_read_reg(const struct device *dev, uint16_t reg, uint32_t *val)
+static int syscon_generic_read_reg(const struct device *dev, uint32_t reg, uint32_t *val)
 {
 	struct syscon_generic_data *data = dev->data;
 	k_spinlock_key_t key;
@@ -110,7 +110,7 @@ static int syscon_generic_read_reg(const struct device *dev, uint16_t reg, uint3
 	return ret;
 }
 
-static int syscon_generic_write_reg(const struct device *dev, uint16_t reg, uint32_t val)
+static int syscon_generic_write_reg(const struct device *dev, uint32_t reg, uint32_t val)
 {
 	struct syscon_generic_data *data = dev->data;
 	k_spinlock_key_t key;
@@ -123,7 +123,7 @@ static int syscon_generic_write_reg(const struct device *dev, uint16_t reg, uint
 	return ret;
 }
 
-static int syscon_generic_update_bits(const struct device *dev, uint16_t reg,
+static int syscon_generic_update_bits(const struct device *dev, uint32_t reg,
 				      uint32_t mask, uint32_t val)
 {
 	struct syscon_generic_data *data = dev->data;
